console: Nintendo 3DS Game Card images: Show the cartridge revision, not the NCCH revision.
There's a separate field indicating the cartridge revision. This isn't
currently documented on 3dbrew.

NOTE: This field might be an leshort, but all known revisions are less
than 256, so we're assuming it's just a byte.

Removed the nintendo-3ds-NCCH common rule, since we don't want to show
the NCCH version for cartridge images.

TODO: Show the System Update partition version?