console: Added Wii U and TNES formats.
Wii U is iNES with "NES\x00" instead of "NES\x1A".

TNES has a completely different 16-byte header, but it's easy enough
to shoehorn in. It also supports FDS, so the FDS info block is used
if the mapper equals 100

Note that TNES uses 8k PRG banks instead of 16k.